Add global variable "block"
Hi,

I never rellay liked my addition of the layoutBlock Function so I replaced it finally with a variable: 

Before: 
```
{{ layoutBlock('someMethod', {foo:bar})
```

After: 
```
{{ block.someMethod(bar) }}
```

I left the old method, so nothing should break, but added a dreprecated warning so that it can be removed some day ;).

Could you merge this as version 1.7? 

Thanks!